Do not throw PlatformNotSupportedException
The `PlatformNotSupportedException` is problematic when the EventLog sink is configured through [Serilog.Settings.Configuration][1], where it's impossible to conditionally enable on Windows and disable on other platforms.

Real scenario: the production app will run on Windows and the appsettings configuration looks like this (using TOML instead of JSON):

```toml
[Serilog.WriteTo.EventLogSink]
Name = 'EventLog'
Args = { source = 'MyApplication', manageEventSource = true, restrictedToMinimumLevel = 'Warning' }
```

When developing on Linux or macOS, it's [impossible to remove][2] the EventLog configuration so it's better for this sink to be ignored since EventLog doesn't exist on these platforms anyway.

Note: this is exactly how the [Notepad sink][3] works and it's a good solution.

[1]: https://github.com/serilog/serilog-settings-configuration
[2]: https://github.com/dotnet/runtime/issues/36543
[3]: https://github.com/serilog-contrib/serilog-sinks-notepad